+++ tests/soroban_testcases/mod.rs
@@ -18,9 +18,9 @@ mod cross_contract_calls;
 mod events;
 mod example_custom_types;
 mod example_events;
+mod example_hello_world;
 mod example_increment_with_pause;
 mod example_pause;
-mod hello_world;
 mod i128_u128;
 mod i256_u256;
 mod integer_width_rounding;
